replace np.int with np.int32

+++ detection/retinaface/rcnn/dataset/ds_utils.py
@@ -4,7 +4,7 @@ import numpy as np
 def unique_boxes(boxes, scale=1.0):
     """ return indices of unique boxes """
     v = np.array([1, 1e3, 1e6, 1e9])
-    hashes = np.round(boxes * scale).dot(v).astype(np.int)
+    hashes = np.round(boxes * scale).dot(v).astype(np.int32)
     _, index = np.unique(hashes, return_index=True)
     return np.sort(index)
